% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextrna_colors.R
\docType{data}
\name{nextrna_colors}
\alias{nextrna_colors}
\title{selected nextrna-themed colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
nextrna_colors
}
\description{
selected nextrna-themed colors
}
\keyword{datasets}
