% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextrna_colors.R
\name{nextrna_cols}
\alias{nextrna_cols}
\title{Function to extract nextrna colors as hex codes}
\usage{
nextrna_cols(...)
}
\arguments{
\item{...}{Character names of nextrna_colors}
}
\description{
Function to extract nextrna colors as hex codes
}
