% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextrna_colors.R
\name{nextrna_pal}
\alias{nextrna_pal}
\title{Return function to interpolate a nextrna color palette}
\usage{
nextrna_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in nextrna_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a nextrna color palette
}
