% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextrna_colors.R
\docType{data}
\name{nextrna_palettes}
\alias{nextrna_palettes}
\title{possible nextrna color palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
nextrna_palettes
}
\description{
possible nextrna color palettes
}
\keyword{datasets}
