% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextrna_colors.R
\name{scale_fill_nextrna}
\alias{scale_fill_nextrna}
\title{Fill scale constructor for nextrna colors}
\usage{
scale_fill_nextrna(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in nextrna_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for nextrna colors
}
